Refactor: divide cli part of code from core operations
